+++ src/Group.php
@@ -48,13 +48,9 @@ final class Group extends Actor
     /**
      * {@inheritdoc}
      */
-    public function equals(Actor $actor)
+    public function equals(Object $actor)
     {
         if (!parent::equals($actor)) {
-            return false;
-        }
-
-        if ('Xabbuh\XApi\Model\Group' !== get_class($actor)) {
             return false;
         }